// frontend/src/sections/cold-chain-equipment/catalogue.generated.ts
export type AssetCatalogueItemsListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    category?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    categoryId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    class?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    classId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    code?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    manufacturer?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    model?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    type?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    typeId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    search?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    subCatalogue?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
  } | null;
  page?: {
    first?: number | null;
    offset?: number | null;
  } | null;
};

export type AssetCatalogueItemsListResult = {
  assetCatalogueItems: ({
  __typename: "AssetCatalogueItemConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  code: string;
  manufacturer: string | null;
  model: string;
  assetCategoryId: string;
  assetTypeId: string;
  assetType: {
  id: string;
  name: string;
} | null;
}>;
});
};

export const AssetCatalogueItemsList = {
  query: "query assetCatalogueItemsList($filter: AssetCatalogueItemFilterInput, $page: PaginationInput) {\n  assetCatalogueItems(filter: $filter, page: $page) {\n    ... on AssetCatalogueItemConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        code\n        manufacturer\n        model\n        assetCategoryId\n        assetTypeId\n        assetType {\n          id\n          name\n        }\n      }\n    }\n  }\n}",
} as TypedDocument<AssetCatalogueItemsListResult, AssetCatalogueItemsListVariables>;

export type AssetPropertiesListVariables = {
  filter?: {
    id?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    key?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    name?: {
    equalTo?: string | null;
    like?: string | null;
  } | null;
    assetClassId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    assetCategoryId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
    assetTypeId?: {
    equalTo?: string | null;
    equalAny?: Array<string> | null;
    notEqualTo?: string | null;
    equalAnyOrNull?: Array<string> | null;
    notEqualAll?: Array<string> | null;
  } | null;
  } | null;
};

export type AssetPropertiesListResult = {
  assetProperties: ({
  __typename: "AssetPropertyConnector";
} & {
  totalCount: number;
  nodes: Array<{
  id: string;
  key: string;
  name: string;
  valueType: "STRING" | "BOOLEAN" | "INTEGER" | "FLOAT" | "DATE";
  allowedValues: string | null;
  assetClassId: string | null;
  assetCategoryId: string | null;
  assetTypeId: string | null;
}>;
});
};

export const AssetPropertiesList = {
  query: "query assetPropertiesList($filter: AssetPropertyFilterInput) {\n  assetProperties(filter: $filter) {\n    ... on AssetPropertyConnector {\n      __typename\n      totalCount\n      nodes {\n        id\n        key\n        name\n        valueType\n        allowedValues\n        assetClassId\n        assetCategoryId\n        assetTypeId\n      }\n    }\n  }\n}",
} as TypedDocument<AssetPropertiesListResult, AssetPropertiesListVariables>;